requirements.txt
aiodns==3.5.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aioquic==1.2.0
aiosignal==1.4.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
attrs==25.3.0
bcrypt==3.2.2
beautifulsoup4==4.14.2
black==25.9.0
blinker==1.9.0
Brotli==1.1.0
cchardet==2.1.7
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==45.0.7
Cython==3.1.4
dataclasses==0.6
debugpy==1.8.17
defusedxml==0.7.1
Django==6.0.2
dnslib==0.9.26
dnspython==2.8.0
docx==0.2.4
driver-manager==0.1.1
et_xmlfile==2.0.0
flake8==7.3.0
Flask==3.1.2
fonttools==4.61.1
fpdf2==2.8.5
frozenlist==1.8.0
greenlet==3.2.4
gunicorn==25.1.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httplib2==0.31.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.3.0
invoke==2.2.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
kaitaistruct==0.10
ldap3==2.9.1
lxml @ file:///home/builder/.termux-build/python-lxml/src
MarkupSafe==3.0.2
mccabe==0.7.0
mitmproxy==12.1.2
mitmproxy_linux==0.12.7
mitmproxy_rs==0.12.7
MouseInfo==0.1.3
msgpack==1.1.1
multidict==6.7.0
mypy_extensions==1.1.0
netifaces==0.11.0
nodeenv==1.9.1
numpy==2.3.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
paramiko==4.0.0
parso==0.8.5
passlib==1.7.4
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
propcache==0.4.1
psycopg2==2.9.11
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycares==4.11.0
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
PyGetWindow==0.0.9
Pygments==2.19.2
pyjnius==1.7.0
pylsqpack==0.3.22
PyMsgBox==2.0.1
PyMySQL==1.1.2
PyNaCl==1.6.0
pynvim==0.6.0
pyOpenSSL==25.1.0
pyparsing==3.2.3
PyPDF2==3.0.1
pyperclip==1.9.0
PyRect==0.2.0
pyright==1.1.406
PyScreeze==1.0.1
PySocks==1.7.1
pysrt==1.1.2
pytest==8.4.2
python-docx==1.2.0
python-dotenv==1.2.1
python-nmap==0.7.1
python3-xlib==0.15
pytokens==0.2.0
pytweening==1.2.0
pywhatkit==5.4
ranger-fm @ file:///home/builder/.termux-build/ranger/src
requests==2.32.5
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
scapy==2.6.1
scipy @ file:///home/builder/.termux-build/python-scipy/src/dist/scipy-1.16.3-cp312-cp312-linux_aarch64.whl#sha256=39d2378d2e195340f3caf1ebc82545c2cd10459bdf805ec9260d488fe92d9bfb
selenium==4.38.0
service-identity==24.2.0
setuptools==80.9.0
simpleaudio==1.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
soupsieve==2.8
sqlparse==0.5.5
termuxgui==0.1.6
tornado==6.5.2
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
urwid==2.6.16
uvloop==0.22.1
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
Werkzeug==3.1.3
wheel==0.45.1
wifidroid==0.0.1
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.22.0
yt-dlp==2025.10.14
zstandard==0.19.0
